lecture 38: create a Django structure from command line
This commit introduces the boilerplate code for a Django project created from the
command 'django-admin startproject personal_portfolio', and renamed the top project
directory name to personal_portfolio_project in order to differentiate it from
the child directory with the same original name. This change is what named
the project in git.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>