lecture 39: create a model
This commit introduces the code for the first model that the application
will have.

To work with images, you maybe will be required to install Pillow
(pip3 install Pillow).

Then to resolve migration problems of the project we need to issue
'python3 manage.py migrate' and to include the new migration that we created,
'python3 manage.py makemigrations'. Then the 0001_initial.py will be created
inside the migrations directory of your app.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>