lecture 40: add admin and Project to admin console
The configuration of the admin console is done via CLI by issuing:

'python3 manage.py createsuperuser'

After configuring it, we can manage the website from the web browser.

The addition of the Project model to the admin.py file under the root directory
of the project allows us to add instances of the model we created on models.py
also from the web browser # Register your models here.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>