lecture 42: add another set of urls
When working with many apps, we can redirect all urls from the root urls.py to a
urls.py file inside the structure of such app.

In this commit we create a urls.py file inside the 'blog' app, so all requests
coming from the localhost:8000/blog url will be forwarded to the blog/urls.py file.

However, in order for that to work, we have to have the same forwarding structure
we already have in place for the personal_portfolio (root) directory, so we create
a templates structure, a urls.py and a views.py files with very similar contents
of the ones present in the root directory.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>