lecture 44: create a blog model and display it
Add the code to create the Blog model class under models.py and register it
in the database (in blog/admin.py), so we can manage it from the admin/
console.

The 0001_initial.py file is created automatically after running the manage.py
createmigrations and migrate commands.

The views.py and the templates files are modified to reflect the contents from
the database on the blogs/ endpoint

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>