lecture 46: pass blog information instead of blog_id
Making use of get_object_or_404, we pass along to the information
related to the object present in the database to detail.html.

The module get_object_or_404 will return the object connected to the id determined in
the 'pk' - primary-key - variable, which, in turn, is blog_id).

In the detail.html page, we display the blog title instead of the blog ID.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>