lecture 43,47: add a date field to blog/models.py
Add a date entry in the database in order to tweak some of the django
formatting in the web page in future lectures.
The 'date' should have been added in lecture 43, but it is being added now.

Note that the two database migration commands should be run after adding the
find in models.py, otherwise the database will not be populated:

$ python3 manage.py makemigrations
$ python3 manage.py migrate

See the changes from the admin console of your website

The change in views.py relates to sorting of the blog entries displayed in the
blogs page. All the sorting options, including slicing of the list of objects
can be found at the Django documentation page at
https://docs.djangoproject.com/en/3.1/topics/db/queries/

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>