lecture 47: make DB entries more readable in the admin page
Adding the __str__() function to the models page we modify the return of the
builtin function by telling it to display whatever information about the
database entries we want, so it is easily accessible from the admin dashboard.

In this case, it is not necessary to make migrations to the DB because it the
changes are not related to the DB tables, but to the dashboard code.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>