lecture 48: add a base.html page with bootstrap templates
Bootstrap [0] contains many html boilerplate code in order to enhance a website.
This commit introduces some formatting (including javascript load triggers) and
a navbar code.

In order to apply that to all html pages in our website, we create a base.html
file with the boilerplate and plug the pages inside the {% block page_content %}
tag.

In every html file in our project, we extend the base and indicate where the hook
tags will be.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>