lecture 55: static and media files in the server
Add the adjustments that are necessary to be made on the server side of the website.
Python anywhere has a straightforward management for static and media files, and
once in place, they can be made effective with a simple `collectstatic` command
from manage.py.

Also add some minor markdown adjustments to the document.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>