lecture 57: add local_settings and requirements.txt
Requirements.txt file was generated with command:

  $ pip3 freeze > requirements.txt

And its function is to install all local libraries and dependencies we have
locally, in the production server.

The local_settings.py was also created and includes the DEBUG_MODE setting and
the ALLOWED_HOSTS as an empty list, so the website can run in our local machine.
However, it is included in .gitgnore because it is only part of the developement
environement and should never go to production.

In order to apply the requirements.txt dependencies on the production server, issue:

  $ pip install -r requirements.txt

Keep in mind that everytime you change anything in the code or database, you need to:

1) install any new dependencies from requirements.txt;
2) migrate all new changes from the database;
3) collect all new static files that might got added.

And reload the page on pythonanywhere.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>